12/6/2011
Part now prints on its side
This allows it to be attached in the vertical direction as opposed to horizontal
The change in design will make a completed printer easier to ship because I can simply loosen the screws attached to the corner brackets and remove the lower z_holder screw.  then rotate the whole z/x assembally 90 degrees.